Inaugurate the test suite
Fortunately the RFC has an example with indexed fields only. Actually
it's _a field_ so the hex dump looks a bit weird.

The test suite is implemented in bourne shell and relies on a new
`hdecode` program that uses the library to print an HPACK frame with
the same format as in the RFC:

   Dynamic Table (after decoding):

   [  1] (s =  54) custom-key: custom-value
   [  2] (s =  53) cache-control: no-cache
   [  3] (s =  57) :authority: www.example.com
         Table size: 164

   Decoded header list:

   :method: GET
   :scheme: https
   :path: /index.html
   :authority: www.example.com
   custom-key: custom-value

There is one difference though: in order to keep `hdecode` stateless
and simple the decoded header list is printed _first_.