Break down the insertion logic further
The FIELD event is no longer handled by HPT_insert. The new entry is
initialized *after* the NAME event was handled so that data from the
table doesn't get kicked out of the table.

It is necessary if the newly inserted field kicks a field out to make
room for itself, but at the same time reuses its name. Smells like even
more dirty pointer arithmetics.