Fix the Travis CI configuration
The `bc` dependency is missing while slightly more exotic programs like
`hexdump` or `xxd` (from vim!) are present in the base image. They are
probably pulled by some language stack.

The `set -e` hack is not good, it seems that travis builds a single
shell script that includes the `after_failure` step too, so it's not
run when there's an error. And displaying the test report really helps
understand build failures, like when `bc` is missing.

Once we go to the build directory, we need to lookup cmake from the
parent directory.