Use a 64-bit integer to store the size of an entry
With HPACK integers limited to 16 bits, the size of a previous entry in
the dynamic table can't fit in 16 bits too. It requires twice 16 bits
for the name and value lengths and the 32B overhead.

For this a signed 64-bit integer should be enough, and with padding both
at the end and the beggining (after the magic number) of the structure,
`struct hpt_entry` should have a portable size of 32B.