Don't install libasan and libubsan on Travis CI
They are advertised as whitelisted [1] but for some reason they are not
available. It makes the apt installation fail, but the build still goes
on and the test suite later fails to find programs it requires.

So instead of excluding gcc(1) from the build matrix for ASAN and UBSAN
usage, it makes more sense to include clang(1) instead. I intially
thought that the bundled gcc 4.6 didn't have *SAN support, but clang(1)
actually bundles those tools.

[1] https://github.com/travis-ci/apt-package-whitelist/blob/master/ubuntu-precise