Insert evicted indexed names in the dynamic table
When a field is inserted in the dynamic table, it may refer to an
already indexed name. But inserting the new field may evict the
referenced name at the same time.

See the insertion flow from RFC 7541 appendix C.5.2.:

   Decoding process:

   48                                      | == Literal indexed ==
                                           |   Indexed name (idx = 8)
                                           |     :status
   03                                      |   Literal value (len = 3)
   3330 37                                 | 307
                                           | - evict: :status: 302
                                           | -> :status: 307

The eviction logically happens after the insertion, so the evicted field
may still be visible during the insertion.

In cashpack the existing entries are moved when a new one is inserted
so the evicted indexed name would be overwritten before being copied.
And because of single allocation design, there's no room in the dynamic
table for buffering of the indexed name. So instead it is incrementally
copied while the rest of the table is moved.

This case was too hard to test using a decoder, but very trivial with an
encoder. It incurs an extra cost but at the same time it should seldom
happen, especially with realistic table sizes far greater than 84 bytes.