An encoder may allocate to the limit
A decoder allocates whatever a table according to the maximum values it
allowed to the peer encoder. An encoder may however allocate a table
lower than the peer decoder allows, according to its own limit. A limit
that it can currently not really set, at least not outside of encoding.