Partial header list encoding
Instead of forcing the user to allocate a complete header list before
encoding it, cashpack now supports incremental encoding of the header
list. Thanks to decoding all the groundwork was already there, and it
brings back some symmetry in the API.

A new `push` statement is available in `hencode` to encode a partial
header list.