Collect decoded strings in the decoding buffer
Instead of using the decoding buffer as a temporary buffer for Huffman
decoding, it is now a persistent region from which decoded strings are
allocated. It is persistent across the decoding of a whole HPACK block,
which can happen in several passes. Passing a different buffer for the
decoding of a CONTINUATION results in undefined behavior, likely an ARG
error.

Strings are written null-terminated to the buffer, which raises a new
BIG error if the whole message doesn't fit. The `hdecode` program was
adapted to accommodate the new rules, but the documentation was left
as-is for now because this will enable more interesting changes.