Decoders no longer send DATA events
Before a field's string is decoded, the position of the buffer where it
will be written is saved. When the string is fully decoded (which may
span several partial blocks) a single NAME or VALUE event is sent.

This allows a first round of simplifications in the library and its
usage.