Generate efficient Huffman lookup tables
Using the same table for both encoding and decoding had the advantage of
keeping things simple, except for the reverse lookup table needed by
encoders.

The new lookup tables trade space for speed. A character will be looked
up in after at most 3 jumps (against 20 today in the worst case) and
requires overall less branching and bit shifting.

With a synthetic benchmark encoding then decoding a random buffer, with
various runs on various sizes, it shows an increase of the decoding
throughput from 94Mb/s to 213Mb/s (+126%) on average.

Refs #3