Add a mechanism to keep tests temp directories
Instead of manually disabling the shell trap that removes the temp
directory, a new shell trap conditionally renames it to avoid the
ineluctable trap on EXIT (except for dash).

The kept directories are cleaned up by relevant make targets, and
they are named after the failing test to find them easily (except
for zsh).