Add interoperability checks in Go
This is my very first program written in Go if the traditional hello
world doesn't count as an actual program. So before I get picky comments
about why this is terrible Go code, it deserves some explanations.

First off, I was discovering the language as I was monkey-writing the
code so I may have not followed TheGoWay(r) and at this point I don't
care yet. It's only a (much shorter) rewrite of the `hdecode` program,
much like the other `ngdecode` program.

So basically, it's only a matter of using Go's native HPACK package
available since Go 1.6, except that for some reason they decided that it
should stay internal for some silly reason. RFC 7541 makes it clear that
this is the "Header Compression for HTTP/2" but doesn't say it should be
used by this protocol only.

Actually, it says:

    Protocols that use HPACK determine the maximum size that the encoder
    is permitted to use for the dynamic table.  In HTTP/2, this value is
    determined by the SETTINGS_HEADER_TABLE_SIZE setting (see
    Section 6.5.2 of [HTTP2]).

Deciding to make it internal makes it harder to prototype a new protocol
on top of HPACK, although Go is a great language for prototyping. If you
really want to use it, you can, but one of my criteria is to be able to
build offline. So I had to trick the `go build` command to pretend I had
previously run `go get` for the HPACK package. I was initially angry at
`gofmt` for not letting me break lines, but I don't why this happened
once and no longer did afterwards (vim integration). So all in all, if
you twist hard enough the Go tooling fits good enough in a make-based
build. It does not play well with Valgrind, which is a bit sad.

What about the language and its standard library? It's actually quite
pleasant to use. Most of the variables benefit from type inference,
error handling is interesting, although I'd love to have a standard
assert package. While I get their point from the FAQ (but disagree), it
would really help at the very least for test programs like this one.
Slices are interesting, closures are very helpful (I wish we had them
in C...) and overall APIs tend to be well designed, except for:

- CamelCaseStartingWithACap? Really?
- the flag package: way too opinionated
- why is syscall.Mmap different than mmap(2)?
- no way to probe the HPACK dynamic table

The most interesting thing with Go's HPACK implementation is that it
agrees with cashpack on the complete test suite! However there's nothing
in the API to tell when a block starts or ends, so it might be possible
to feed junk to a decoder and get away with it.