Dear Travis CI, I hate old Ubuntu releases
I couldn't find a way to list clang's warnings, and the error I've seen
come from the -W option. For GCC it's a deprecated alias to -Wextra, but
it doesn't look like it in clang. And the warnings I get, for which
errors are now disabled still trigger on ye olde clang 3.4 on Ubuntu
Trusty.

Let's hope it fixes it for good.