Shameless synthetic coverage
This takes the cashpack library to a 100% coverage according to lcov. Of
course it doesn't mean that everything is covered, but that's another
story.

There are two lines not covered by the library: both of them are WRONG
statements that should NEVER occur. Accordingly, the two branches that
lead to those statements are also not covered.