Play cat and mouse with Go 1.7
The HPACK code being a private thing for some reason beyond me, it is by
nature subject to change without prior notice. I hope it won't happen
again. It can always be worked around by running `go get` and overriding
the GOPATH.

An interesting thing with the GOROOT was that I was using gcc-go instead
of the native Go toolchain. The former doesn't ship the sources and has
a different root directory. This is now meant to work with native Go and
otherwise there's always the GOPATH to patch things up together.