Extend syntastic integration
Weaken the _SYNTASTIC macro, tell the compiler what to comply with, and
give hints to find other include directories. Generating the options
could be done at configure time to keep them in sync but the syntastic
author is not interested in falling back to the working directory. So
checking a file from $srcdir would not work in a VPATH setup.