Polish the distributed header file
This is a terrible hack that nobody should need to do. We can ask cpp(1)
to keep comments but it may then drop comments from implicitly-included
files that we never asked in the first place.

It works by pre-processing an empty file, and counting the resulting
number of lines. That's how many lines we then want to remove from the
pre-processed `hpack.h` file. Keeping the comments around also means
keeping the copyright notice when distributing the code.

Since the point is to distribute a self-contained single-file header, it
also means having to work around the copyright notices from the included
tables.

The `hpack.h` file is now created iff all transformations succeed.