Be pedantic, de-anonymize all unions
Anonymous unions are very convenient, that's the one C11 extension I'm
looking forward to see available everywhere. It doesn't clutter the code
that much, and maybe it's a sign that data structures have too way much
nesting.

This is a patch mostly mechanical. Now that the -pedantic flag is always
set for GCC or clang, it will be hard to miss C99 rules.