Don't use cashpack types to keep track of state
This is in `cashdumb`, which is far from ideal currently, especially as
manual material. Allocations are tracked in dumb data structures, and
cashpack input is managed independently. This is overall less hackish.

This is more apparent when it comes to cleanup code. There's a clear
segregation between state management and calls to `hpack_clean_field`.