Wrap flag detection inside a shell function
It saves up about 800 lines of generated shell code, and has the great
advantage of making this part of the configure script slightly wee more
readable. It looks like this:

    cashpack_check_flags() {
        for _flag
        do
            [... unreadable autoconf code ...]
        done
    }

    if test "$cross_compiling" = no; then :
       CASHPACK_CFLAGS=

       cashpack_check_flags -Werror -Wall -W [... more FreeBSD flags ...]
       cashpack_check_flags -Wextra -Wmissing-declarations [... moar ...]
       cashpack_check_flags -errwarn=%all -errtags=yes
       cashpack_check_flags -Wsparse-all -Wsparse-error
       cashpack_check_flags -pedantic -std=c99 -D_POSIX_C_SOURCE=200809L

       CFLAGS="$CASHPACK_CFLAGS $CFLAGS"
    fi

This is overall performing better and easier to review in a dist
archive. And... I'm omitting the numerous empty lines that make
the all thing slightly less readable than I'd hoped, but it's also
important to keep the m4 code readable too.