Turn Clang's -Walmost-everything-under-the-sun on
This change was tricky for more than one reason. First, the way flags
are detected by cashpack was conflicting. Despite having already used
`AC_RUN_IFELSE` directly and removed the broken solution available in
the autoconf archive, it turns out that defined macros would still be
prepended to the compiled source code and trigger a -Weverything warn
that would fail because of -Werror. Fiddling with autoconf is always a
bit scary, but moving a file seems to do the trick. Second, some of the
code constructs (like the WRONG macro) would trigger a warning, only if
assertions are enabled. Third, system includes may do more silly things
too — like recursive macro definitions — and it may be a problem on
other systems too. It can still be worked around by setting C flags at
either configure or build time. Fourth, a number of warnings needed to
be fixed along with enabling the warning. Fifth, it turns out that it may
not be a good idea to enhance your paranoia when it comes to examples
or tests. The formers need to be kept as simple as possible, and for the
latter it would prevent intentional misuse of the library. Last, flags
need to be added last, otherwise -Weverything breaks the detection of
sanitizers.

That being said, fixing those warnings uncovered a missing check in
field encoding. The length of name and value fields should fit in a
16bit unsigned integer. Code coverage dropping below 100% will be the
reminder, and a test case will land later.