Move the build madness to a shell script
In order to ship a self-contained `hpack.h`, includes need preprocessing.
However according to the autoconf manual:

> Macro: AC_PROG_CPP
>
>     Set output variable CPP to a command that runs the C preprocessor.
>     If ‘$CC -E’ doesn't work, /lib/cpp is used. It is only portable to
>     run CPP on files with a .c extension.

And FreeBSD with Clang were happy to demonstrate how not portable it was.
This was actually the one failure too much, so instead of maintaining
cryptic make rules it now lives in a standalone script.

In addition, the double-hash (##) hack to preserve defines is now gone,
only to be replaced by a safer hack. It now produces a valid header with
GCC's, Clang's and pcc's preprocessors. The best results are obtained
using GCC's standalone cpp(1) command.

Honorary mention to cpp(1) on OSX that now spits incorrect code, forcing
the introduction of a new double-hash hack.