Teach decoders how to skip large messages
The BIG error is now recoverable by removing anything decoded before the
current field, and moving the current field at the beginning of the now
empty buffer. It is still possible to see a single field not fit in the
whole decoding buffer and an SKP error is introduced for that case.

After an overflow all bets are off on the consumer side, so a decoder
will stop invoking the callback once a message is deemed too big.