Coverage for `hpack_search`
Back to 100% coverage. In order to efficiently test the binary search,
the existing facilities weren't making things easy. Especially with so
many branches in the code, the test data needed to be large enough.

Since a binary search was supposed to help performance, instead of the
usual abstract tests written in shell, this one is a micro benchmark.
Taking the linear search as a baseline with very unscientific measures,
the two optimizations yield an improvement in execution time:

- ~22% faster for pseudo-headers only
- ~70% faster for all static entries