Don't print resource usage on MacOS
That was supposed to be a portable solution and it looks like it checks
out in MacOS according to the manual but on Travis CI it doesn't. Also
amazingly getrusage(2) is in POSIX but I couldn't find a publication
describing `struct rusage`. Cherry on the cake, `struct timeval` in
MacOS uses and int instead of a long int for the microseconds (which is
fair to say, _enough_) while Linux uses long int for both seconds and
microseconds, probably because it doesn't make a difference in structure
size because of alignment.

So for now, ignore MacOS, depending on tests on other platform it may
become Linux-only reporting. *sigh*