Update Travis CI configuration for Ubuntu Trusty
It has gradually become the default, and thanks to a CRON job the moment
cashpack made the switch to Trusty it blew up.

The automake version shipped uses a parallel test harness by default,
and even using the serial test harness no longer prints the test report
directly in the output. The VERBOSE environment variable is now set,
and the new requirement is automake 1.12+.

The Clang version is not compatible with GCC's and lcov's versions, and
code coverage can only be collected using GCC now. On the plus side the
LLVM update was able to find a new leak in the test suite with ASAN.

The memcheck execution takes more than 10 minutes for the longest test,
so the travis_wait command is now needed. The upside is that Valgrind
and automake are now both recent-enough to drop their workarounds.

More importantly, the packages that were missing in the Trusty whitelist
when Precise went EOL all seem to be present now.