Travis CI maintenance
Put sanitizers configuration in the global environment for clarity,
neuter rst2man for osx jobs (in the previous occurrence of the build
system it was implicit,) merge asan and ubsan jobs, revive the clang
coverage job.

The next step is to hopefully get the reason why asan started failing
out of the blue a couple weeks ago.