Keep sanitizers Clang-only on Travis CI
GCC is still too old for that, and now the game is: what changed that
would make ASAN fail with EPERM?

    AddressSanitizer Init done
    Could not attach to thread <main> (errno 1).
    Failed suspending threads.
    LeakSanitizer has encountered a fatal error.

It looks like the CAP_SYS_PTRACE capability is gone, maybe because of a
docker upgrade?

https://docs.travis-ci.com/user/build-environment-updates/2017-12-12/

Since there are already Valgrind jobs, it's probably simpler to disable
the leak sanitizer for now. While at it, fix the clang coverage job.