What goes around comes around
The Go tool chain is horrendous, and revisiting cashpack has always been
met with a new creative way to fail the build for Go checks in the test
suite, wasting my time as I try to make it work again without making it
too intrusive.

The godecode program fits in a single source file and does not need all
the bells and whistles of a large scale project, but Go modules and the
generally irresponsible dependency management of the tool chain insist
strongly on asking cashpack to download the HPACK code that also happens
to be present "vendored" in the standard library, alongside the rest of
the tool chain.

This is absurd.

What is also absurd is hiding the h2 and HPACK implementations, despite
the very nature of the Go build process that requires all sources to be
present at compile time. Why even create a pseudo-external dependency
for the standard library?

Maybe Go doesn't have such a nice package management if requiring all
sources at build time also means that some interfaces you don't want to
export are de-facto exported and all you can do is hiding them behind
extra vendoring hoops.

One could argue the benefits of the go.sum file for cashpack's SBOM, but
if a trusted Go tool chain was already installed, and if the HPACK code
was already "vendored" alongside the standard library, then downloading
the HPACK code and giving me a checksum is actually weakening my supply
chain until I go and audit what was downloaded.

This design is broken, but when in Rome...

This time, `go build` returns with a vengeance, creating a module on the
fly and downloading the latest version of the HPACK implementation, and
since cashpack is not a Go project, isolated from the default GOPATH.

Let's see how long this one will last.