Track consumed integer bytes
This is not tracking the length of the the integer, but rather the
length of input consumed in the last decoding round. This is in effect
the length of the integer if it fits entirely in an entire or partial
HPACK block, when the integer does not span over two partial blocks.